requirements.txt
Django
dj-database-url
dj-static
gunicorn
Unipath
python-decouple
Pillow
Markdown
django-heroku
bleach
psycopg2
whitenoise
djangorestframework
psycopg2-binary
dj3-cloudinary-storage
django-storages
boto3
django-filter==2.4.0
django-allauth
django-cors-headers
django-humanize
django-resized
django-import-export
pandas
matplotlib